Move algorithms into separate files
